Keep rust targets from getting rebuilt.
I think what's happening here is that the `.lock` file is being updated
at the end of the `cargo build` process, after the binary has been
written. Whatever the case is, one of the dependencies is getting its
mtime updated to a later time, so the solution is the same regardless
of who the offender is.